requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.3
dataclasses-json==0.5.14
distlib==0.3.7
dnspython==2.4.2
et-xmlfile==1.1.0
faiss-cpu==1.7.4
fastapi==0.101.0
filelock==3.12.2
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
idna==3.4
langchain==0.0.271
langsmith==0.0.26
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
openai==0.27.8
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
pandas-stubs==2.0.3.230814
pdfminer.six==20221105
pdfplumber==0.10.2
Pillow==10.0.0
platformdirs==3.10.0
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pymongo==4.5.0
PyPDF2==3.0.1
pypdfium2==4.18.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
regex==2023.8.8
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
tenacity==8.2.3
tiktoken==0.4.0
tqdm==4.66.1
types-pytz==2023.3.0.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.2
virtualenv==20.24.2
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
